requirements-demo.txt
numpy>=1.24
pandas>=2.0
opencv-python>=4.8
Pillow>=10.0
matplotlib>=3.8
folium>=0.15
streamlit>=1.31
streamlit-folium>=0.18
